// web/src/lib/search/interfaces.ts
export const FlowType = {
  SEARCH: "search",
  QUESTION_ANSWER: "question-answer",
};
export type FlowType = (typeof FlowType)[keyof typeof FlowType];
export const SearchType = {
  SEMANTIC: "semantic",
  KEYWORD: "keyword",
  AUTOMATIC: "automatic",
  INTERNET: "internet",
};
export type SearchType = (typeof SearchType)[keyof typeof SearchType];

export interface ToolResponse {
  id?: string | null;
  response?: any;
}
export interface ExtendedToolResponse extends ToolResponse {
  level: number;
  level_question_num: number;
}

export enum StreamStopReason {
  CONTEXT_LENGTH = "CONTEXT_LENGTH",
  CANCELLED = "CANCELLED",
}

export interface StreamStopInfo {
  stop_reason: StreamStopReason;
  level?: number;
  level_question_num?: number;
  stream_type?: "sub_answer" | "sub_questions" | "main_answer";
}

export interface ErrorMessagePacket {
  error: string;
}

export interface Quote {
  quote: string;
  document_id: string;
  link: string | null;
  source_type: ValidSources;
  blurb: string;
  semantic_identifier: string;
}

export interface QuotesInfoPacket {
  quotes: Quote[];
}
export interface MinimalVirchowDocument {
  document_id: string;
  semantic_identifier: string | null;
}

export interface VirchowDocument extends MinimalVirchowDocument {
  link: string;
  source_type: ValidSources;
  blurb: string;
  boost: number;
  hidden: boolean;
  score: number;
  chunk_ind: number;
  match_highlights: string[];
  metadata: { [key: string]: string };
  updated_at: string | null;
  db_doc_id?: number;
  is_internet: boolean;
  validationState?: null | "good" | "bad";
}

export interface LoadedVirchowDocument extends VirchowDocument {
  icon: React.FC<{ size?: number; className?: string }>;
}

export interface SearchVirchowDocument extends VirchowDocument {
  is_relevant: boolean;
  relevance_explanation: string;
}

export interface FilteredVirchowDocument extends VirchowDocument {
  included: boolean;
}
export interface DocumentInfoPacket {
  top_documents: VirchowDocument[];
  predicted_flow: FlowType | null;
  predicted_search: SearchType | null;
  time_cutoff: string | null;
  favor_recent: boolean;
}
// ...
export interface Filters {
  source_type: string[] | null;
  document_set: string[] | null;
  time_cutoff: Date | null;
}